feat: embed をオプションからカスタマイズできるように修正
