+++ packages/zenn-markdown-html/package.json
@@ -53,6 +53,7 @@
     "markdown-it-inline-comments": "^1.0.1",
     "markdown-it-link-attributes": "^4.0.1",
     "markdown-it-task-lists": "^2.1.1",
+    "mermaid": "^9.3.0",
     "prismjs": "^1.27.0",
     "sanitize-html": "^2.7.2"
   },
