+++ packages/zenn-markdown-html/src/types.ts
@@ -1,6 +1,9 @@
+import type { EmbedGeneratorList } from './embed';
+
 /**
  * Markdown 変換時のオプション型
  */
-export interface MarkdownOptions {
+export type MarkdownOptions = {
   embedOrigin?: string;
-}
+  customEmbed?: Partial<EmbedGeneratorList>;
+};
