+++ packages/zenn-markdown-html/src/utils/embed-helper.ts
@@ -1,173 +1,10 @@
-import { escapeHtml } from 'markdown-it/lib/common/utils';
 import { MarkdownOptions } from '../types';
-import {
-  isGistUrl,
-  isTweetUrl,
-  isGithubUrl,
-  isStackblitzUrl,
-  isCodesandboxUrl,
-  isCodepenUrl,
-  isJsfiddleUrl,
-  isBlueprintUEUrl,
-  isFigmaUrl,
-  isYoutubeUrl,
-} from './url-matcher';
-import { extractYoutubeVideoParameters } from './url-matcher';
-
-/* 埋め込み要素の種別 */
-export type EmbedType =
-  | 'youtube'
-  | 'slideshare'
-  | 'speakerdeck'
-  | 'jsfiddle'
-  | 'codepen'
-  | 'codesandbox'
-  | 'stackblitz'
-  | 'tweet'
-  | 'blueprintue'
-  | 'figma'
-  | 'card'
-  | 'gist'
-  | 'github'
-  | 'mermaid';
-
-/** embedサーバーで表示する埋め込み要素の種別 */
-export type EmbedServerType = Extract<
-  EmbedType,
-  'tweet' | 'card' | 'mermaid' | 'github' | 'gist'
->;
-
-type EmbedGenerator = (token: string, options?: MarkdownOptions) => string;
-
-/** 渡された文字列から埋め込み要素のHTMLを生成する関数をまとめたオブジェクト */
-const embedGenerators: Record<EmbedType, EmbedGenerator> = {
-  youtube(str) {
-    const params = extractYoutubeVideoParameters(str) || { videoId: str };
-
-    if (!params.videoId.match(/^[a-zA-Z0-9_-]+$/)) {
-      return 'YouTubeのvideoIDが不正です';
-    }
-
-    const escapedVideoId = escapeHtml(params.videoId);
-    const time = Math.min(Number(params.start || 0), 48 * 60 * 60); // 48時間以内
-    const startQuery = time ? `?start=${time}` : '';
-
-    return `<span class="embed-block embed-youtube"><iframe src="https://www.youtube-nocookie.com/embed/${escapedVideoId}${startQuery}" allow="accelerometer; clipboard-write; encrypted-media; gyroscope; picture-in-picture" allowfullscreen loading="lazy"></iframe></span>`;
-  },
-  slideshare(key) {
-    if (!key?.match(/^[a-zA-Z0-9_-]+$/)) {
-      return 'Slide Shareのkeyが不正です';
-    }
-    return `<span class="embed-block embed-slideshare"><iframe src="https://www.slideshare.net/slideshow/embed_code/key/${escapeHtml(
-      key
-    )}" scrolling="no" allowfullscreen loading="lazy"></iframe></span>`;
-  },
-  speakerdeck(key) {
-    if (!key?.match(/^[a-zA-Z0-9_-]+$/)) {
-      return 'Speaker Deckのkeyが不正です';
-    }
-    return `<span class="embed-block embed-speakerdeck"><iframe src="https://speakerdeck.com/player/${escapeHtml(
-      key
-    )}" scrolling="no" allowfullscreen allow="encrypted-media" loading="lazy"></iframe></span>`;
-  },
-  jsfiddle(str) {
-    if (!isJsfiddleUrl(str)) {
-      return 'jsfiddleのURLが不正です';
-    }
-    // URLを~/embedded/とする
-    // ※ すでにembeddedもしくはembedが含まれるURLが入力されている場合は、そのままURLを使用する。
-    let url = str;
-    if (!url.includes('embed')) {
-      url = url.endsWith('/') ? `${url}embedded/` : `${url}/embedded/`;
-    }
-    return `<span class="embed-block embed-jsfiddle"><iframe src="${sanitizeEmbedToken(
-      url
-    )}" scrolling="no" frameborder="no" loading="lazy"></iframe></span>`;
-  },
-  codepen(str) {
-    if (!isCodepenUrl(str)) {
-      return 'CodePenのURLが不正です';
-    }
-    const url = new URL(str.replace('/pen/', '/embed/'));
-    url.searchParams.set('embed-version', '2');
-    return `<span class="embed-block embed-codepen"><iframe src="${sanitizeEmbedToken(
-      url.toString()
-    )}" scrolling="no" frameborder="no" loading="lazy"></iframe></span>`;
-  },
-  codesandbox(str) {
-    if (!isCodesandboxUrl(str)) {
-      return '「https://codesandbox.io/embed/」から始まる正しいURLを入力してください';
-    }
-    return `<span class="embed-block embed-codesandbox"><iframe src="${sanitizeEmbedToken(
-      str
-    )}" style="width:100%;height:500px;border:none;overflow:hidden;" allow="accelerometer; ambient-light-sensor; camera; encrypted-media; geolocation; gyroscope; hid; microphone; midi; payment; usb; vr; xr-spatial-tracking" loading="lazy" sandbox="allow-modals allow-forms allow-popups allow-scripts allow-same-origin"></iframe></span>`;
-  },
-  stackblitz(str) {
-    if (!isStackblitzUrl(str)) {
-      return 'StackBlitzのembed用のURLを指定してください';
-    }
-    return `<span class="embed-block embed-stackblitz"><iframe src="${sanitizeEmbedToken(
-      str
-    )}" scrolling="no" frameborder="no" loading="lazy"></iframe></span>`;
-  },
-  tweet(str, options) {
-    if (!isTweetUrl(str)) return 'ツイートページのURLを指定してください';
-    return generateEmbedServerIframe('tweet', str, options);
-  },
-  blueprintue(str) {
-    if (!isBlueprintUEUrl(str))
-      return '「https://blueprintue.com/render/」から始まる正しいURLを指定してください';
-    return `<span class="embed-block embed-blueprintue"><iframe src="${sanitizeEmbedToken(
-      str
-    )}" width="100%" style="aspect-ratio: 16/9" scrolling="no" frameborder="no" loading="lazy" allowfullscreen></iframe></span>`;
-  },
-  figma(str: string) {
-    if (!isFigmaUrl(str))
-      return 'ファイルまたはプロトタイプのFigma URLを指定してください';
-    return `<span class="embed-block embed-figma"><iframe src="https://www.figma.com/embed?embed_host=zenn&url=${sanitizeEmbedToken(
-      str
-    )}" width="100%" style="aspect-ratio: 16/9" scrolling="no" frameborder="no" loading="lazy" allowfullscreen></iframe></span>`;
-  },
-  card(str, options) {
-    if (!isValidHttpUrl(str)) return 'URLが不正です';
-    return generateEmbedServerIframe('card', str, options);
-  },
-  gist(str, options) {
-    if (!isGistUrl(str)) return 'GitHub GistのページURLを指定してください';
-    /**
-     * gistのURL は
-     * - https://gist.github.com/foo/bar.json
-     * - https://gist.github.com/foo/bar.json?file=example.js
-     * のような形式
-     */
-    return generateEmbedServerIframe('gist', str, options);
-  },
-  github(str, options) {
-    if (!isGithubUrl(str)) {
-      return 'GitHub のファイルURLまたはパーマリンクを指定してください';
-    }
-    return generateEmbedServerIframe('github', str, options);
-  },
-  mermaid(src, options) {
-    return generateEmbedServerIframe('mermaid', src, options);
-  },
-} as const;
-
-const embedKeys = Object.keys(embedGenerators) as EmbedType[];
+import { embedKeys, EmbedServerType, EmbedType } from '../embed';
+import { isTweetUrl, isGithubUrl, isYoutubeUrl } from './url-matcher';
 
 /** 渡された文字列をサニタイズする */
-function sanitizeEmbedToken(str: string): string {
+export function sanitizeEmbedToken(str: string): string {
   return str.replace(/"/g, '%22');
-}
-
-/** URL文字列か判定する */
-function isValidHttpUrl(str: string) {
-  try {
-    const url = new URL(str);
-    return url.protocol === 'http:' || url.protocol === 'https:';
-  } catch (_) {
-    return false;
-  }
 }
 
 /** `EmbedType`か判定する */
@@ -203,11 +40,11 @@ export const validateEmbedToken = (
 export function generateEmbedServerIframe(
   type: EmbedServerType,
   src: string,
-  options?: MarkdownOptions
+  embedOrigin: string
 ): string {
   const origin = (() => {
     try {
-      return new URL(options?.embedOrigin || '').origin;
+      return new URL(embedOrigin).origin;
     } catch {
       return void 0;
     }
@@ -235,7 +72,9 @@ export const generateEmbedHTML = (
   options?: MarkdownOptions
 ): string => {
   const { isValid, message } = validateEmbedToken(str, type);
-  return isValid ? embedGenerators[type](str, options) : message;
+  const generator = options?.customEmbed?.[type];
+
+  return isValid ? generator?.(str, options) || str : message;
 };
 
 /** Linkifyな埋め込み要素のHTML生成する */
@@ -244,16 +83,18 @@ export const generateLinkifyEmbedHTML = (
   options?: MarkdownOptions
 ): string => {
   const { isValid, message: msg } = validateEmbedToken(url);
+  const generators = options?.customEmbed;
+
+  if (!generators) return url;
 
   if (isTweetUrl(url))
-    return isValid ? embedGenerators.tweet(url, options) : msg;
+    return isValid ? generators.tweet?.(url, options) || url : msg;
 
   if (isYoutubeUrl(url))
-    return isValid ? embedGenerators.youtube(url, options) : msg;
+    return isValid ? generators.youtube?.(url, options) || url : msg;
 
   // GitHub は URL が長くなりやすいためバリデーション(`validateEmbedToken`)から外す
-  if (isGithubUrl(url))
-    return generateEmbedServerIframe('github', url, options);
+  if (isGithubUrl(url)) return generators.github?.(url, options) || url;
 
-  return generateEmbedServerIframe('card', url, options);
+  return generators.card?.(url, options) || url;
 };
